requirements.txt
python-telegram-bot==21.6
fastapi==0.115.4
uvicorn[standard]==0.32.0
aiosqlite==0.20.0
pydantic==2.9.2
web3==6.11.1
httpx==0.27.2
